requirements.txt
requests
pandas
beautifulsoup4
psycopg2-binary
tzdata
